RUBY-1501 Refine new request agent agent attributes
This is another implementation that adds the new request agent
agent attributes.  It incurs one additional allocation over the
previous commit, but is a cleaner solution overall.  We can evaluate
whether or not this is worth the additional allocation.

Here are the performance stats against an agent that does not collect
the additional request attributes.  Note again, that 8 object allocations
are the result of calling Rack::Request#host.

mwear:ruby_agent mwear$ ./test/performance/script/runner -n test_basic_middleware_stack -C
1 tests, 0 failures, 7.592214 s total
+--------------------------------------------+-----------+-----------+-------+---------------+--------------+--------------+
| name                                       | before    | after     | delta | allocs_before | allocs_after | allocs_delta |
|--------------------------------------------+-----------+-----------+-------+---------------+--------------+--------------|
| RackMiddleware#test_basic_middleware_stack | 429.58 µs | 433.41 µs | +0.9% |           244 |          253 |        +3.7% |
+--------------------------------------------+-----------+-----------+-------+---------------+--------------+--------------+