RUBY-1494 Modify URLMap test to work for Puma only environment
This modifies the url_map_test to work for a Puma only environment.
To do this we had to drop the dependency on Rack::Test.  Since we
were only using the `get` helper method I implemented a barebones
one inline. If this ultimately becomes a problem we can revisit
how we want to test the Puma only env.