RUBY-1494 Fix disable_rack and disable_rack_urlmap flags
When adding the disable flags disable_puma_rack and disable_puma_rack_urlmap
I noticed that the disable_rack_urlmap flag is not used and that the
rack url map instrumentation checks the disable_rack flag.  I also noticed
that the rack instrumentation does not check the disable_rack flag.  This
commit adds a check for disable_rack to the Rack instrumentation and changes
the check for installing rack urlmap instrumentation from checking disable_rack
to disable_rack_urlmap.