RUBY-1521 Conditionally expect response.content_length based on Rails Version
We test using both ActionController::IntegrationTest and ActionDispatch::IntegrationTest
and they differ in how they set the headers for response.content_type.
ActionController::IntegrationTest sets this value in the headers portion of the rack
response whereas ActionController::TestCase sets the value after the fact. Therefore
we need to conditionally expect the response.content_length to be set during tests.