RUBY-1520 Extract synthetics event responsibilities into SyntheticsEventAggregator
This commit introduces code that was extracted from the TransactionEventAggregator
to handle synthetics events.  There are some interactions between the TEA and SEA
that still need to be restored, but that will be done through the Transaction
class. This commit also includes basic tests for the SEA which will be augmented
in coming commits.