RUBY-1520 Remove samples accessor from TransactionEventAggregator
The samples accessor on the TEA, while synchronized, provides unsafe
access to the data in the underlying buffer.  Unsafe in that it returns
a copy of the samples that could be mutated by a receiver without a mutex.
The method was only used for testing purposes. Tests have been updated to access the
samples by harvesting them.