RUBY-1520 Extract TransactionEvent class into TransactionEventPrimitive module
Extracted the TransactionEvent class into a module for a couple of reasons.
The motivating factor was that the TransctionEvent class did not really have
any behavior. It's basically just a collection of methods that transforms
the transaction payload into the expected collector primitive. A secondary
motivation is that using a module saves us an allocation per transaction.
I also moved a large number of tests from the TransactionEventAggregatorTest
to the TransactionEventPrimitive test as they are testing the primitive
generated and did not really have anything to do with the aggregator.