Removing the link to the app role
Deployments that don't use the :app role are failing because of the fact that this job is fixed to the :app role. Why is there even a link to the app role? What does it have to do with a role? We just need a single deployment message to NewRelic.