RUBY-1520 Only create transaction events if the buffer admits them
When the transaction event aggregrator's buffer is full a transaction event
may not be admitted depending on the reservoir sampling algorithm.  This PR
updates the TransactionEventAggregator to accept a block that creates the event
which will only be called if the event is going to be admitted to the buffer.