RUBY-1520 Replace frozen constants for symbols in TransactionEventPrimitive
The use of frozen constants in TransactionEventPrimitive was holdover from
its origins in the TransctionEventAggregrator. We replace them with symbols
as they have the same performance qualities with slightly better readability.
Also, the error event hash construction uses symbols, and for reasons of
consistency we should pick one strategy to use in both places.