RUBY-1520 Don't under count samples merged from the PipeService
Datacontainers can have data merged from one of two sources: a failed harvest,
or via the PipeService. To ensure that we have correct metadata we need to
ensure that we don't double data that has been merged from a failed harvest
and that we don't inadvertently under count data merged via the PipeService.