RUBY-1520 Use frozen string as key for timestamp on transaction events
caef318cf7bf6750a3fb4fa5c286874ed7dd6f8d moved from using frozen strings
to symbols for transaction event keys. This is fine for all but one case,
the timestamp, which is used by the SyntheticsEventAggregator to admit
or reject new samples. In the case of a failed harvest and subsequent
merge the symbol :timestamp would be converted to a string 'timestamp'.
We could check for either or the string or symbol version in the
SyntheticsEventAggregator, we could always use strings, or we could revert
caef318cf7bf6750a3fb4fa5c286874ed7dd6f8d which would use strings for all
keys. In this commit I chose the second option as a compromise.