Change how we specify a database adapter for JRuby
The activerecord-jdbc-adapter readme reccommends ommitting the 'jdbc'
prefix from your database.yml's adapter setting on Rails >= 3.0.
However, that does not appear to be true for all minor 3.x releases.
Additionally, the :platform groups are confusing in conjunction with
logical controls in the Envfile, so we will avoid using platform groups.