Fix multiverse active record mysql dependencies
We would like to avoid including the activerecord-mysql2-adapter
whenever possible, as it is simply an extract of the mysql2 gem,
and it is included in active record in versions 3.1 and above.
We will instead use the mysql adapter for ruby 1.9.3 and below.
We will use the mysql2 adapter for rubies above 1.9.3, and the
activerecord-jdbcmysql-adapter for all versions of JRuby.