Fix datamapper password masking on 187/ree
On 1.8.7 / ree, datamapper would return the datastore URI
with its parameters in an arbitrary order. This violates
our previous observation of the password parameter appearing
consistently in the middle of the parameters string on
rubies that maintain hash element order. This commit allows
our password scrubber to detect the edge case when the
password key-value pair is shuffled to the end of the URI
query string.