RUBY-1536 Update test with new expectation based on Rails 5 behavior
Rails has a number of different behaviors as to what it does when a
template is not found and hence, the agent has a number reactions.
The end result is that the agent will not record a metric for
view rendering, which is acceptable as this is the same behavior
as it had in Rails 3.