RUBY-1552 Fix new getMore MGI issue
With the MongoCommandSubscriber the value of event.command.values.first
will be the collection name for all commands other than getMore. We were
bit by this before and added special case handling for the 'getMore' command.
When running the most recent mongo driver (v 2.2) with the most recent release
of the database (v 3.2.4) the event.command_name is the symbol :getMore, previously
it was the string 'getMore'. This commit updates the instrumentation to handle
the getMore command as either a string or symbol.