RUBY-1562 Compress based on size, rename config option
The spec for simple compression indicated payloads should be compressed
when using simple compression. After some discussion it was determined
that they should be compressed based on previous agent conventions and
the main point is that individual json documents within the payload
should remain uncompressed. The simple_compression config option was
also mislabled as simple_encoding. That has been fixed.