RUBY-1563 Don't let one connect failure on Resque startup spoil the fun
In the context of Resque, we call manual_start explicitly from the Resque master
process, with the sync_startup option set to true. This causes the agent to
attempt to connect to the collector synchronously before any jobs are processed.

If this connection attempt fails, however, we would previously never retry, and
would leave the agent in the disconnected state, losing instrumentation about
all jobs processed by that Resque instance.

With this change, we'll give it one shot synchronously, but if that fails we'll
allow the connection retries to continue in the background. This means we might
miss some jobs if the first connection attempt fails, but at least we'll pick up
monitoring again once we successfully connect.

This also means that if you happen to get a failure on your first call to
manual_start where you pass in :keep_retrying => false, you won't be stuck
permanently in a disconnected state.