RUBY-1566 Add ability to record metrics to AbstractSegment
This commit adds a template method record_metrics as well as a
method to access a metric_cache to do the recording. Depending
on whether this segment is part of a transaction the method cache
will either be the local TransactionMetrics or the StatsEngine
on the Agent::Agent. I had to alias a couple of methods in stats
engine so that TransactionMetrics and the StatsEngine have the
same interface for recording metrics. I would prefer not to do
that, but it's the cleanest solution for now.