RUBY-1566 Introduce machinery to build new segment api atop the TracedMethodStack
This commit lays the foundations for us to build the new segment API on top
of the legacy tracer internals. Temporarily we will push segments onto the
traced method stack along with the usual TracedMethodFrames. This allows to
implement the segment API along with the metric naming, recording and timing
responsiblities of the segment class. Once we have introduced the segment API
and update instrumentation to use it we can replace legacy tracer internals
with those from the new tracer prototype.