RUBY-1566 Rewrite trace_execution_scoped using new segment API
Thie commit rewrites trace_execution_scoped to use the new segment
API. This is a good test to show that the machinery is in place and
working properly because the majority of our instrumentation will
at some point call trace_execution_scoped. There were a handful
of changes that had to be made to the method_tracer_test for tests
that were asserting scopes using the MockScopeListener class. The
behavior of trace_execution_scoped was unchanged by this commit,
but the implementation is different, especially in the case where
a segment is not bound to a transaction. Previously, traced_execution_scope
would call transaction.notice_push / pop_frame regardless of whether there
was an active transaction. If there wasn't an active transaction then those
methods would no-op. Using the segment API we bypass that code path altoghether
if there isn't an active transaction. The tests were written in such a way that
they were making assertions based on notice_push / pop_frame being called in
trace_execution_scoped. Those tests were updated so that an active transaction
would be started by using the in_transaction helper to force hitting the old
code path.