RUBY-1566 Rescue unfinishable segments
This commit adds a rescue block in segment finish, which I don't
particularly like, but without it the bare multiverse test suite
fails. Apparently it is possible for trace_execution_scoped to be
executed via a method_tracer without the agent having been initialized
yet. In that case the NewRelic::Agent.instance.stats_engine is nil
(so is NewRelic::Agent.instance). We were protected by a rescue in
traced_execution_scoped_footer previously. Adding this rescue to
the finish method gives us the same protection. While this is less
than ideal, it seems that is necessary for now. See:
test/multiverse/suites/bare/standalone_instrumentation_test.rb for
more details.