RUBY-1567 Add scoped_metric_for to Datastores::MetricHelper
In order to implement the DatastoreSegment (based on the API spec) we
need to more fine grained access to the to metrics that are generated
from the MetricHelper. This commit adds a scoped_metric_for method to
the helper that will return the name of the most specific metric
for the product, operation and collection in question.