RUBY-1576 Fix instrumentation for Sequel 4.35.0
In version 4.35.0 Sequel renamed and changed the signature of
log_yield to log_connection_yield. log_yield is the method that we
base our instrumentation off of for the Sequel extension. This commit
updates the instrumentation to work with the new method and signature
and adds version 4.35.0 to the multiverse suite.