RUBY-1568 Add ignore! method to segment to disable tracing
The Sequel instrumentation makes use of NewRelic::Agent.disable_all_tracing
which sets a flag to turn off creation of tt nodes and metrics. The segment
already has a flag to turn off metric recording, but does not have one to
to omit it from a trace. This commit adds an ignore flag which does both:
omits the segment from the trace and does not record metrics. This also updates
the Tracing module to ignore segments when the is_execution_traced? flag is false.