Don't test DelayedJob with ActiveRecord3 and Ruby 2.3+
ActiveRecord 3.1 introduced a comparison without doing guarding.
Ruby 2.3 no longer rescues exceptions of `<=>` and starts
returning `false` instead of `nil`. This comparison was later
fixed in AR3.2 to return nil, but as DelayedJob itself only
tests Rails 3 with up to Ruby 2.1, we can drop this combination
of testing with Ruby 2.3.