RUBY-1584 Simplify segment ignoring
I ran into another issue where instrumentation uses a
NewRelic::Agent.disable_all_tracing block. My previous solution did
not take into account when a disable_all_tracing block was executed
outside of an active transaction. While revisiting ignoring a segment
I realized that this is not something the segment really needs to be
concerned about it and that we can solve in the tracing
module. If we want to ignore a segment we want to make sure that it
is not part of the transaction trace and that we do not record metrics
for that segment. We do, however, want to return a segment so that
our code doesn't always have to guard around whether or not it has
a reference to a segment. This commit updates the tracing module
so that if tracing is disabled, we do not add the segment to the
current transaction and we toggle metric recording off on the segment.
This is cleaner than the previous solution that added an additional
ignore flag to the segment.