RUBY-1584 Handle case where operation and collection raises in Mongo instrumentation
There is some careful programming around exceptions being raised in the
metrics_for method of the MetricTranslator. Presumably these same
issues could arise when trying to determine the operation and collection, since
that method was extracted from metrics_for. I'm unsure if these are real
possibilities or just imagined and coded around, but to be safe I've preserved
this behavior and made the instrumentation tolerant of a segment not being
created when an exception is raised determining the operation and collection.