RUBY-1584 Remove metrics_for method from Mongo::MetricTranslator
We only need to know the collection and operation in order to create
our datastore segment and it will use the Datastores::MetricHelper
to generate the metrics. In this commit I've removed the unused
metrics_for method from the Mongo::MetricTranslator, but had to
update the metric_translator_test to test that the proper
operation and collection were determined rather than testing
the proper metrics generated. There were several tests that
were deleted as they were test behavior of the
Datastores::MetricHelper and no longer applied to determining
the operation and collection.