Preemptively require an appropriate version of json for older rubies
Version 2.0 of the json requires Ruby 2.0 or newer. We have logic
to pin the json version when it is specified in the Gemfile, but
sometimes the json library is pulled in by libraries as a secondary
dependency. To work around this situation we preemptively add json
to the gemfile with an appropriate version for rubies < 2.0. This
is not a great solution, but will hopefully allow CI to pass until
we can do something more permanent. Ideally this PR to bundler:
https://github.com/bundler/bundler/pull/4650 that will make it
respect Ruby versions will be merged. If that doesn't happen in
a timely manner we should come up with a better solution for
multiverse.