RUBY-1581 Extract product_operation_collection_for from Datastores::MetricHelper
To create a datastore segment we need to be able to determine the product, operation,
and collection for the segment. That logic was bundled into to the metrics_for method.
This commit extracts it into its own method and updates some of the tests for the
metric_helper to expect product, operation, and collection rather than metrics.