RUBY-1582 Use datastore segment in ActiveRecordSubscriber
This commit updates the ActiveRecordSubscriber to use the datastore
segment. However, it does not use the API and the instrumentation is
still far from ideal. This because we are working around two issues.
One issue is the traced_method_stack on the current transaction tracer.
It expects frames to be popped off in the same order they were pushed on.
The ActiveRecord subscriber does not always finish events in the same
order they were started. This issue will be resolved as we replace the
tracer internals.

The other problem that we are working around is that the new segment
types do not yet carry out all of responsiblities that they need to.
In the future the notice_sql method will move from the
TransactionSampler to the datastore segment. For the time being we
use the ActiveRecord event to store the necessary information to
track slow sql. We will utlimately move that responsiblity to the
segment and should be able to do away with the ActiveRecordEvent
completely.