fall back to |-joined versions for edge cases
grape 0.16 changed the `Route` API and `#route_version` became
`#version`, which now may return an Array under certain default
behavior. this commit adds tests for those cases when env['api.version']
is not set, and joins the version strings with '|' before inclusion in
the transaction name in order to mimic pre-0.16 behavior.