RUBY-1596 Make transaction_state available to all segments
Ultimately we would like to do away with transaction state by moving most of
the functionality to the transction itself, but for now there are a number of
questions we rely on the state to answer. This commit adds a method that
will make state available to all segments regardless of whether there is an
active transaction or not.