RUBY-1596 Add notice_sql_statement to TransctionSampler
This commit adds notice_sql_statement, a variation of notice_sql,
that takes a simplified parameter list which includes a
Database::Statement object. This will be called from the
DatastoreSegment which will be responsible for created the
the Statement object. This will allow us to share the same
Statement object between the SqlSampler and the TransactionSampler
as currently each of the samplers create separate (yet identical)
instances of the Database::Statement.