RUBY-1596 Add notice_sql method to DatastoreSegment
This commit adds an _notice_sql method to the datastore segment. The
leading underscore is to indicate that this is not meant for public
use. We will provide a simplified public version that does not allow
users to specify an explain plan. We are in the process of doing away
with the stack and tree based tracer and migrating to a segment only
tracer. As part of this transition we need to move the node parameters
to the segment. In the case of a datastore segment the node parameters
will include slow sql. Moving notice_sql to the datastore segment gives
us the API we want, but since we are still using the old tracer it
is just a middleman for the data. This will not be the case when we
complete the segment based tracer.

Moving this method here allows us to create a single database statement
object that is shared by both the TransactionSampler and SqlSampler.
Previously both the samplers created their own identical instances
of the database statement.