RUBY-1596 Ensure database statement does not mutate sql argument
Moving the notice_sql method to the datastore segment gives us
more flexibility on when the method can be called. Before it was
always called after a query has run. Now it can potentially be
called before. We dup the incoming sql so that it isn't mutated
before it's executed.