RUBY-1596 Dup the Database::Statement
One of the goals I had while moving the notice_sql method to the Datastore
segment was to make it possible for the TransactionSampler and SqlSampler
to both share the same database statement. This is still an achieveable goal,
but I have found one issue that makes it more challening than it should be.
The issue is that SqlSampler nils out the statement's explainer on harvest.
That introduces a race condition where the the TransactionSampler would
not to be able to run an explain plan. The reason the SqlSampler nils out
the explainer that is uses Marshall.dump to send data over a Pipe for IPC
in the Resque instrumentation. We can still fix this issue and make it so
that both the SqlSampler and Transaction sampler can both use the the same
statment, but it is going to take more work than we currently have time for.
I'll leave this for a future improvement.