RUBY-1596 Ensure sql is only dup'd once
We want to dup an sql statement so that we are free to modify it and will
not cause problems for a library executing the statement, but we want to
ensure the dup happens only once. The correctly_encoded helper method will
dup a string if it needs to change its encoding. This commit adds logic to
the capture_query method to dup the incoming sql if it's not modified by
correctly_encoded.