Cobble together perf test for AR Subscriber
This commit includes a cobbled together perf test for the AR Subscriber.
Portions are taken from the AR subscriber unit test and Rails source
since perf tests do not have a great way to bring in dependencies.
I'm commiting this test because it's useful, not because it's
particularly well written. For the record, these are the results of
running this test against this revision compared to 3.16.0:

```
+---------------------------------------------------+-----------+-----------+-------+---------------+--------------+--------------+
| name                                              | before    | after     | delta | allocs_before | allocs_after | allocs_delta |
|---------------------------------------------------+-----------+-----------+-------+---------------+--------------+--------------|
| ActiveRecordSubscriberTest#test_subscriber_in_txn | 150.19 µs | 154.82 µs | +3.1% |            95 |           98 |        +3.2% |
+---------------------------------------------------+-----------+-----------+-------+---------------+--------------+--------------+
```