RUBY-1602 Fix AbstractSegment#inspect
Inspecting a segment results in a very long string when it is part
of a transaction as it has a reference to the transaction_state, which
has a reference to the traced_method_stack which then proceeds to print
a representation of each node. This commit omits the problematic
instance variables from when inspect is called.