RUBY-1602 Do not push segments on the stack outside of transaction for AR Subscriber
Due to idiosyncrasies of the both the AR Subscriber and the traced_method_stack
the subscriber bypasses Transaction.start_datastore_segment and pushes segments
directly onto the traced_method_stack. This commit fixes a bug where outside of
an active transaction the subscriber will push a segment on the stack, but never
pops it off.