RUBY-1286 Add trace_only option to notice_error
Normally, calling notice_error will record the trace, increment the
error count, and consider the transaction as failing for Apdex
purposes.

With this PR, you can pass the :trace_only option to perform just the
first step (recording the error) without affecting the error rate or
the succeeded/failed status for Apdex calculations.

This change enables us to log detailed backtrace information for
errors that are important enough to fix, but not urgent enough to page
us on a high error rate.