RUBY-1629 Separate InstanceIdentifier.for method into host and port_path_or_id methods
This commit is part of a refactor to split the instance identifier into
two parts, host and path_port_or_id. This commit extracts methods that
will provide the host or path_port_and_id individually and updates
the related tests. The InstanceIdentifier.for method will remain until
later in the refactor where it will be able to be safely removed.