RUBY-1690 Basic perf test for external segments
This commit introduces a perf test for external segments. The test
launches a local server in a child process and uses Net::HTTP to
make a web request inside of a transaction. The allocation numbers
are accurate, the timing numbers are going to be variable since
they are dependent on the response times of the local server. Here
the results of a comparision against 3.17.2:

mwear:ruby_agent mwear$ be ./test/performance/script/runner -n test_external_request -C
1 tests, 0 failures, 7.979427 s total
+---------------------------------------+---------+---------+-------+---------------+--------------+--------------+
| name                                  | before  | after   | delta | allocs_before | allocs_after | allocs_delta |
|---------------------------------------+---------+---------+-------+---------------+--------------+--------------|
| ExternalSegment#test_external_request | 3.15 ms | 3.23 ms | +2.8% |           593 |          586 |        -1.2% |
+---------------------------------------+---------+---------+-------+---------------+--------------+--------------+