RUBY-1569 External CAT perf test
This commit adds a test for external cat. Similarly to the external request
perf test, the allocations will be accurate, but execution time will vary
as the test makes requests to a webserver running in a child process. Here
are the results of a comparision against 3.17.2 of the agent.

mwear:ruby_agent mwear$ be ./test/performance/script/runner -n test_external_cat_request -C
1 tests, 0 failures, 7.959484 s total
+-------------------------------------------+---------+---------+-------+---------------+--------------+--------------+
| name                                      | before  | after   | delta | allocs_before | allocs_after | allocs_delta |
|-------------------------------------------+---------+---------+-------+---------------+--------------+--------------|
| ExternalSegment#test_external_cat_request | 4.17 ms | 3.98 ms | -4.5% |           664 |          625 |        -5.9% |
+-------------------------------------------+---------+---------+-------+---------------+--------------+--------------+