RUBY-1730 notice_sql on segment for Sequel
This commit updates the Sequel instrumentation to use
DatastoreSegment#_notice_sql instead of the soon to be removed
TransactionSampler#notice_sql method. When using the `all` method
on a Sequel Model, slow sqls will not reflect that name instead
of `select` which was the previous behavior. This brings the
behavior in line with what our ActiveRecord instrumentation does.