RUBY-1747 Introduce TraceBuilder
This commit introduces a TraceBuilder module that will build a trace
from a flat list of segments using the existing Trace and TraceNode
classes. We may want to clean up the Trace and TraceNode classes as there
are some optimizations we can make to them given the way we are building traces
or we may want to consider building the tree out of the segments themselves.