RUBY-1747 Build trace from flat list of segments at end of Transaction
In this commit we transition from building transaction traces on the fly,
to building them at the end of the transaction. This is a big step along
the path of building transaction traces on demand. In future commits
we can begin to dismantle the old machinery that was used to build traces
on the fly.