RUBY-1747 Improve TraceBuilder algorithm
This optimizes trace building to make fewer object allocations.
We now reuse the arrays allocated when making the relationship
map rather than creating new ones to store TraceNode children.

Here are the perf tests for us building the trace out of the
flat list of segments, vs us building the trace on the fly using
the old machinery.

mwear:ruby_agent mwear$ be ./test/performance/script/runner -n test_basic_middleware_stack -C
1 tests, 0 failures, 7.50442 s total
+--------------------------------------------+-----------+-----------+--------+---------------+--------------+--------------+
| name                                       | before    | after     | delta  | allocs_before | allocs_after | allocs_delta |
|--------------------------------------------+-----------+-----------+--------+---------------+--------------+--------------|
| RackMiddleware#test_basic_middleware_stack | 447.81 µs | 399.36 µs | -10.8% |           284 |          298 |        +4.9% |
+--------------------------------------------+-----------+-----------+--------+---------------+--------------+--------------+

We are 10% faster with 5% more object allocations. We should
be able to shave off a few more object allocations. This is a big
win so far as we will ultimately only bulid the traces we need,
rather than a trace for each transaction.