RUBY-1747 Change TraceNode constructor, make params lazy to save allocations
There was a faux new2 contructor on TraceNode used by TraceBuilder that was
transitionary. We've now adopted it as the offical constructor for TraceNode.
Additionally, we made params lazy to save on object allocations. In the previous
commit I included perf tests that showed we had a 10% increase in speed when
deferring trace creation, but a slight increase in object allocations. These
changes give us the same speed increase, but with 0 new object allocations!

+--------------------------------------------+-----------+-----------+--------+---------------+--------------+--------------+
| name                                       | before    | after     | delta  | allocs_before | allocs_after | allocs_delta |
|--------------------------------------------+-----------+-----------+--------+---------------+--------------+--------------|
| RackMiddleware#test_basic_middleware_stack | 447.81 µs | 399.70 µs | -10.7% |           284 |          284 |         0.0% |
+--------------------------------------------+-----------+-----------+--------+---------------+--------------+--------------+