RUBY-1742 Use haml-rails for Rails 4+
Since Rails 4, the preferred way to have the haml template renderer
installed is by using haml-rails. Some of the multiverse environments
were getting around this by locking down to an older version of haml,
however environments with the unqualified haml references started failing
with the release of haml 5. This change updates our rails 4+ environments
to use haml 5 and haml-rails for all environments.