RUBY-1742 Handle cached queries for ActiveRecord 5.1 and add to Multiverse
Changes were made to the ActiveRecordSubscriber payload in Rails 5.1.
In previous versions of Rails the :name would be set to CACHED to indicate
a cached query. For Rails 5.1 a key :cached will be present in the payload.
This commit dynamically defines a method, somewhat grossly, to test if a
query is cached across the supported Rails versions.