RUBY-1760 Remove rescues from start_message_broker_segment methods
The rescues were added before the Bunny instrumentation was written,
and while their intentions are good, the right place handle issues
with segment creation is in the instrumentation. We covered our
bases in the Bunny instrumentation, so those rescues can be removed.
As they were they made the API some challenging to use. For example,
if you were missing an argument when trying to use it, it would
silently swallow the ArgumentError and you'd have to spelunk through
the logs to see why instrumentation failed.