RUBY-1760 add start_time to segment constructors
during instrumentation of Bunny::Queue#pop, it became evident that we
would need to set a start time on the consume segment that was prior to
actually creating the segment. this change allows start_timet to be
passed into any segment, and only sets it during AbstractSegment#start
if it is not already set.