RUBY-1747 Post-rebase test fixups
This commit fixes two tests that were failing after rebasing with
dev. The reason the tests fail is that they start segments that
aren't finished by the end of the transaction. That in turn was
triggering an exception while building the trace. Ultimately, we
need to make sure trace building is robust enough to detect this
situation and handle it gracefully.