RUBY-1747 Handle unfinished segments, with caveats
This commit allows us to handle segments that weren't finished
properly. With the framestack this situation would have resulted
in a number of lost trace nodes due to its pop-up-the-stack
behavior. We are able to do slightly better with our stackless
implementation. With the stackless tracer, we will forcibly finish
segments that are not complete at the end of the transaction. However,
since the parents nodes will have already completed, the exclusive
time will be incorrect. This might be something we can fix in the
future. At any rate, this will result in a segment that is effectively
running concurrently with many others in our trace. Since we do not
yet have proper accounting for exclusive time for concurrent segments,
the exclusive time will still be off. This is a failure case, and one
where we used to fail much less gracefully. This is a step in the right
direction. We will never be able to have entirely accurate numbers
when this situation arises. The goal is to trace what we can accurately,
and log enough information to indicate that the data might be incorrect.