RUBY-1747 Defer trace creation until harvest
This is the goal of the branch of work, deferring trace creation until
harvest. The buffers that store transactions now need to work in mixed
mode. That is, some of the items will be transaction objects, others will
be traces. For the most part the items should be transactions, but in the
event of a failed harvest we will merge in trace objects. In order
to facilitate this a handful of methods needed to be added to the Transaction
class to give it the same interface as Trace. Another oddity of this change
is that it is possible for a Transaction to be admitted to more than one buffer.
One case where this can happen is if you have a synthetics request that is also
the slowest transaction. In this case we will end up building two traces since
each buffer holds on to a raw transaction and converts it into a trace at the
last minute. There are several ways we can solve this problem, we'll settle on
one in an upcoming commit.

Here are the initial perf results:

mwear:ruby_agent mwear$ be ./test/performance/script/runner -n test_basic_middleware_stack -C
1 tests, 0 failures, 7.516735 s total
+--------------------------------------------+-----------+-----------+--------+---------------+--------------+--------------+
| name                                       | before    | after     | delta  | allocs_before | allocs_after | allocs_delta |
|--------------------------------------------+-----------+-----------+--------+---------------+--------------+--------------|
| RackMiddleware#test_basic_middleware_stack | 448.47 µs | 400.10 µs | -10.8% |           286 |          250 |       -12.6% |
+--------------------------------------------+-----------+-----------+--------+---------------+--------------+--------------+