RUBY-1747 Avoid edge case by moving trace creation to TransactionSampler
This fixes the situation described in the last commit, where we could end
up creating a trace twice if, for example, you had a synthetics request
that was also your slowest transaction. Previously, each buffer would
convert transaction to traces independently. This change moves trace
creation outside of the buffers to the TransactionSampler. Now we can
unique the collection of transactions before creating traces, and avoid
building a trace twice.