RUBY-1733 Remove special case timing code when record_metrics is false
There is unusual special case timing code that has been preserved
that we are removing in this commit. If there is an active transaction
and a segment is started with record_metrics set to false, then that
segment inherits the children time of its grandchildren, rather than the
duration of its children. This behavior is odd at best and going
to be increasingly hard to preserve moving forward with async time
calculations. I've searched through the agent and RPM site to see
if this special case is every hit, and it is not. The agent
will set record_metrics to false only when outside of an active
transaction, and RPM site does not use the option at all.