RUBY-1733 Avoid second iteration over segments in finalize
This commit reworks some code so that we avoid iterating over the
segment list twice during finalize. This gives us a performance gain
at the expense of having incorrect exclusive time for parents where
a child segment is not properly finished during a trasaction. We
consider this an error condition and do not expect it to happen,
but we want to be resilient if it does. We can deal with the
improper exclusive time in return for better performance on all
transactions where segments are properly finished.